pyproject.toml
[project]
name = "aqilas"
version = "0.1.0"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
description = "Library for aqilas sms"
authors = [{ name = "Mohamed ZEBA", email = "parice02@hotmail.com" }]
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.8"
license = "MIT"
license-files = ["LICENSE"]
dependencies = ["requests>=2.31.0"]
keywords = ["burkina aqilas sms", "aqilas sms", "burkina sms"]

[project.urls]
Homepage = "https://github.com/mzeba/aqilas-sms"
Issues = "https://github.com/mzeba/aqilas-sms/issues"
Repository = "https://github.com/mzeba/aqilas-sms.git"

[build-system]
requires = ["setuptools>=77.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = ["pytest", "python-dotenv"]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
